\name{accuracy-class}
\Rdversion{1.1}
\docType{class}
\alias{accuracy-class}
\alias{plot,accuracy-method}

\title{Class "accuracy" }
\description{These objects hold the accuracy computation for the classification of chromosomes into inverted or non-inverted population for increasing levels of BIC thresholds. }
\section{Objects from the Class}{
\code{accuracy} is generated by calls to \code{accBic(inversionList, mem, classFile, nsub, npoints, geno, wROI)}, which is a method for the class \code{inversionList}
}
\section{Slots}{
  \describe{
    \item{\code{out}:}{Object of class \code{"matrix"} }
    matrix with three columns storing the range of BIC thresholds, probability of   inversion within the population and accuracy of the classification. 
  }
}
\section{Methods}{
  \describe{
    \item{plot}{\code{signature(x = "accuracy",w="character")}: plots accuracy Vs BIC for \code{w="a"}, or frequency of inversion Vs BIC for \code{w="f"} }
	 }
}
\examples{
data(ac)
ac
plot(ac)
}


\author{ Alejandro Caceres \email{acaceres@creal.cat} }
\note{version R 2.10.1}
\seealso{\code{\link{accBic}}, \code{\linkS4class{inversionList}} }
\keyword{classes}
