% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversions.R
\name{C_to_K}
\alias{C_to_K}
\title{Celsius conversion to Kelvin}
\usage{
C_to_K(C_temp)
}
\arguments{
\item{C_temp}{The temperature in degrees Celsius}
}
\value{
The temperature in degrees Kelvin
}
\description{
Convert degrees Celsius temperatures to degrees Kelvin
}
\examples{
temp1 <- C_to_K(22);
temp2 <- C_to_K( c(-2, 12, 23) );
}
