% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_capacidad_pago.R
\name{calcular_capacidad_pago}
\alias{calcular_capacidad_pago}
\title{monthly payment capacity}
\usage{
calcular_capacidad_pago(
  ingresos,
  deudas,
  tasa,
  plazo_meses,
  entrada,
  gastos_hipotecarios,
  comisiones
)
}
\arguments{
\item{comisiones}{}
}
\description{
monthly payment capacity
}
\examples{
calcular_capacidad_pago(3000,500,4.2,300,15000,7.5,7.5)
#262500
}
