% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_deuda_final.R
\name{calcular_deuda_final}
\alias{calcular_deuda_final}
\title{Calcula el coste de toda la hipoteca
Calculate the total cost of the mortgage.}
\usage{
calcular_deuda_final(tasa, cuota, prestamo, plazo_meses, fecha_inicial)
}
\arguments{
\item{tasa}{Anual tax}

\item{cuota}{Monthly quota to pay}

\item{prestamo}{Import mortgage loan (prestamo)}

\item{plazo_meses}{mortgage term in months}

\item{fecha_inicial}{Start of mortgage}
}
\value{
Return total cost of the mortgage
}
\description{
Calcula el coste de toda la hipoteca
Calculate the total cost of the mortgage.
}
\examples{
\dontrun{
calcular_deuda_final(4.2,1200,200000,300,"2024-12-31")
}

}
