% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_intereses_pagados.R
\name{calcular_intereses_pagados}
\alias{calcular_intereses_pagados}
\title{Calcula los intereses que se van a pagar durante toda la hipoteca
Calculate the interest to be paid over the entire mortgage.}
\usage{
calcular_intereses_pagados(tasa, cuota, prestamo, plazo_meses, fecha_inicial)
}
\arguments{
\item{tasa}{Anual tax}

\item{cuota}{Monthly quota to pay}

\item{prestamo}{Import mortgage loan (prestamo)}

\item{plazo_meses}{mortgage term in months}

\item{fecha_inicial}{Start of mortgage}

\item{tasacion}{Value of the building, flat, house}
}
\value{
Return interest to be paid over the entire mortgag
}
\description{
Calcula los intereses que se van a pagar durante toda la hipoteca
Calculate the interest to be paid over the entire mortgage.
}
\examples{
\dontrun{
recolecta_datos_de_resultado(4.2,1200,200000,300,"2024-12-31")
}

}
