% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carga_bbdd_inmuebles.R
\name{carga_bbdd_inmuebles}
\alias{carga_bbdd_inmuebles}
\title{Tittle Load database from sql file or load files if is first time ever you start.}
\usage{
carga_bbdd_inmuebles()
}
\arguments{
\item{provincias}{Dataframe con datos de las provincias que ademas contiene las regiones}

\item{municipios}{Dataframe con datos de los municipios}

\item{idealista_data}{Datos de los inmuebles que se cargaron en el pasado de Idealista. BBDD interna para explotacion de historico}
}
\value{
c(provincias, municipios,idealista_data))Return 3 tables with provincias, municipios and idealista_data
}
\description{
Tittle Load database from sql file or load files if is first time ever you start.
}
\examples{
carga_bbdd_inmuebles()

}
