% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crea_tabla_amortizacion.R
\name{crea_tabla_amortizacion}
\alias{crea_tabla_amortizacion}
\title{Create amortizacion table with basic infomration accordind input valuees}
\usage{
crea_tabla_amortizacion(tasa, cuota, prestamo, plazo_meses, fecha_inicial)
}
\arguments{
\item{tasa}{Annual tax}

\item{cuota}{Monthly quota to pay}

\item{prestamo}{Import mortgage loan (prestamo)}

\item{plazo_meses}{mortgage term in months}

\item{fecha_inicial}{Start of mortgage}
}
\value{
Return table with amortizacion. there are 5 columns
}
\description{
Create amortizacion table with basic infomration accordind input valuees
}
\examples{
\dontrun{
crea_tabla_amortizacion(4.2,1200,300,200000,"2024-12-31")
}

}
