% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crea_tabla_roi.R
\name{crea_tabla_roi}
\alias{crea_tabla_roi}
\title{Create ROI table}
\usage{
crea_tabla_roi(tasa, cuota, prestamo, plazo_meses, fecha_inicial)
}
\arguments{
\item{tasa}{Anual tax}

\item{cuota}{Monthly quota to pay}

\item{prestamo}{Import mortgage loan (prestamo)}

\item{plazo_meses}{mortgage term in months}

\item{fecha_inicial}{Start of mortgage}
}
\value{
Return table with amortizacion. there are 5 columns
}
\description{
Create ROI table
}
\examples{
\dontrun{
crea_tabla_roi(4.2,1200,300,200000,"2024-12-31")
}
}
