\name{HaploCode-class}
\Rdversion{1.1}
\docType{class}
\alias{HaploCode-class}
\alias{initialize,HaploCode-method}
\alias{show,HaploCode-method}

\title{Class "HaploCode" }
\description{The object stores the haplotype coding for each of the candidate brake-points. It is typically generated by calling the function \code{codeHaplo} on a \code{GenoDat} object.
}
\section{Objects from the Class}{
Objects can be created by calls of \code{ codeHaplo(objectGenoDat, blockSize, minAllele, saveRes = TRUE, file = NULL, ROI)}

}
\section{Slots}{
  \describe{
    \item{\code{haploCode}:}{Object of class \code{"matrix"} Haplotype coding into decimal integers}
    \item{\code{blockSize}:}{Object of class \code{"numeric"} block size of SNP used to identify the haplotype of each candidate brake-point. }
    \item{\code{minAllele}:}{Object of class \code{"numeric"} minimum allele above which candidate brake-points are considered}
  }
}
\section{Methods}{
  \describe{
    \item{initialize}{\code{signature(.Object = "HaploCode")}: Internal, users should use \code{codeHaplo}}
    \item{show}{\code{signature(object = "HaploCode")}: shows HaploCode}
	 }
}
\author{Alejandro Caceres \email{acaceres@creal.cat}}

\seealso{
\code{\link{codeHaplo}},     \code{\linkS4class{GenoDat}}
}
\examples{
data(gDat)
hapCode <-codeHaplo(gDat,blockSize=3,minAllele=0.3,saveRes=FALSE)
hapCode
}
\keyword{classes}
