% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversions.R
\name{K_to_F}
\alias{K_to_F}
\title{Kelvin conversion to Fahrenheit}
\usage{
K_to_F(K_temp)
}
\arguments{
\item{K_temp}{The temperature in degrees Celsius}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
Convert degrees Celsius temperatures to degrees Fahrenheit
}
\examples{
temp1 <- K_to_F(22);
temp2 <- K_to_F( c(-2, 12, 23) );
}
