% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recolecta_datos_de_alquiler.R
\name{recolecta_datos_de_alquiler}
\alias{recolecta_datos_de_alquiler}
\title{Calculate profits based on the expenses and income from a rental}
\usage{
recolecta_datos_de_alquiler(
  gastos_alquiler,
  mantenimientos_alquiler,
  ingresos_alquiler
)
}
\arguments{
\item{gastos_alquiler}{#Annual expenses}

\item{mantenimientos_alquiler}{#Annual expenses}

\item{ingresos_alquiler}{#Annual earnings}
}
\value{
Return table with all the information and final profit
}
\description{
Calculate profits based on the expenses and income from a rental
}
\examples{
gastos_alquiler <-1000
mantenimientos_alquiler <- 1000
ingresos_alquiler <-15000
\usage{
  recolecta_datos_de_alquiler(gastos_alquiler,mantenimientos_alquiler,ingresos_alquiler)
}
recolecta_datos_de_alquiler(1000,1000,15000)
# Output: 1083.33
}
