% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recolecta_datos_de_resultado.R
\name{recolecta_datos_de_resultado}
\alias{recolecta_datos_de_resultado}
\title{Recopila los datos mas importantes y lo devuelve en formato HTML con colores
Create output result with all the core values}
\usage{
recolecta_datos_de_resultado(
  tasacion,
  tasa,
  cuota,
  prestamo,
  plazo_meses,
  fecha_inicial,
  ingresos,
  deuda
)
}
\arguments{
\item{tasacion}{Value of the building, flat, house}

\item{tasa}{Anual tax}

\item{cuota}{Monthly quota to pay}

\item{prestamo}{Import mortgage loan (prestamo)}

\item{plazo_meses}{mortgage term in months}

\item{fecha_inicial}{Start of mortgage}

\item{ingresos}{earning monthly}

\item{deuda}{Debt monthly}

\item{deuda_final}{Final value of all the mortgage}

\item{intereses_pagados}{All the interests to pay during all the mortgage}
}
\value{
Return table with amortizacion. there are 5 columns
}
\description{
Recopila los datos mas importantes y lo devuelve en formato HTML con colores
Create output result with all the core values
}
\examples{
\dontrun{
recolecta_datos_de_resultado(180000,4.2,1200,200000,300,"2024-12-31",4000,500)
}

}
