\name{scan-class}
\Rdversion{1.1}
\docType{class}
\alias{scan-class}
\alias{plot,scan-method}
\alias{show,scan-method}

\title{Class "scan"}
\description{Results from scanning the genome with the inversion model for trial segments of fixed window size. 
}
\section{Objects from the Class}{
\code{scan} objects are typically generated by callings to the constructor function \code{scanInv}
}
\section{Slots}{
  \describe{
    \item{\code{leftBP}:}{Object of class \code{"matrix"} Left brake-point coordinates (right SNP)}
    \item{\code{rightBP}:}{Object of class \code{"matrix"} Right brake-point coordinates (right SNP)}
    \item{\code{leftBP2}:}{Object of class \code{"matrix"} Left brake-point coordinates  (left SNP)}
    \item{\code{rightBP2}:}{Object of class \code{"matrix"}  Right brake-point coordinates  (right SNP) }
    \item{\code{LogLike}:}{Object of class \code{"matrix"} Log-likelihood ratio for each trial segment }
    \item{\code{prob}:}{Object of class \code{"matrix"} probability of no-inversion for each trial segment }
    \item{\code{ent}:}{Object of class \code{"matrix"} entropy for each trial segment }
    \item{\code{entTh}:}{Object of class \code{"matrix"} entropy threshold for each trial segment}
    \item{\code{bic}:}{Object of class \code{"matrix"} BIC for each trial segment }
    \item{\code{window}:}{Object of class \code{"numeric"} window size }
  }
}
\section{Methods}{
  \describe{
    \item{getInv}{\code{signature(object = "scan")}: gets scan results into a matrix }
    \item{getROIs}{\code{signature(object = "scan")}: get regions of interest, overlapping trial segments with significant BIC}
    \item{listInv}{\code{signature(object = "scan")}: determines the inversion sequence for each ROI}
    \item{plot}{\code{signature(x = "scan")}: plots \code{scan} results, set option \code{which=c("bic","log","prob","ent")} to plot BIC, log-likelihood ratio, probability of no inversion or entropy; and \code{thBic=0} to plot segments with BIC greater than 0 }
    \item{show}{\code{signature(object = "scan")}: shows scan results for each ROI}
	 }
}
\author{Alejandro Caceres \email{acaceres@creal.cat} }

\seealso{ \code{\link{listInv}}, \code{\linkS4class{HaploCode}}, \code{\link{getInv}}, \code{\link{getROIs}}
}
\examples{
data(scanRes)
scanRes
plot(scanRes,which="bic",thBic=0)
}
\keyword{classes}
