% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_capacidad_pago.R
\name{calcular_capacidad_pago}
\alias{calcular_capacidad_pago}
\title{monthly payment capacity}
\usage{
calcular_capacidad_pago(
  ingresos,
  deudas,
  tasa,
  plazo_meses,
  entrada,
  gastos_hipotecarios,
  comisiones
)
}
\arguments{
\item{ingresos}{earning monthly}

\item{deudas}{Debt monthly}

\item{tasa}{Anual tax}

\item{plazo_meses}{mortgage term in months}

\item{entrada}{Debt monthly}

\item{gastos_hipotecarios}{#Son los de gestion}

\item{comisiones}{#Comisiones bancarias}
}
\value{
capacidad_pago #Devuelve la capacidad de pago
}
\description{
monthly payment capacity
}
