% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recolecta_datos_de_resultado.R
\name{recolecta_datos_de_resultado}
\alias{recolecta_datos_de_resultado}
\title{Tittle Create output result with all the core values}
\usage{
recolecta_datos_de_resultado(
  tasacion,
  tasa,
  cuota,
  prestamo,
  plazo_meses,
  fecha_inicial,
  ingresos,
  deuda,
  entrada,
  entrada_minima
)
}
\arguments{
\item{tasacion}{Value of the building, flat, house}

\item{tasa}{Anual tax}

\item{cuota}{Monthly quota to pay}

\item{prestamo}{Import mortgage loan (prestamo)}

\item{plazo_meses}{Mortgage term in months}

\item{fecha_inicial}{Start of mortgage}

\item{ingresos}{Earning monthly}

\item{deuda}{Debt monthly}

\item{entrada}{Down payment contributed to cover initial expenses and 90 per 100 of the sale price.}

\item{entrada_minima}{Minimum Down payment contributed to cover initial expenses}
}
\value{
tabla_resultado Return table with amortizacion. There are 5 columns
}
\description{
Tittle Create output result with all the core values
}
\examples{
tasacion <- 200000
tasa <- 90
cuota <- 1500
prestamo <- 180000
plazo_meses <- 360
fecha_inicial <- as.Date("2023-01-01")
ingresos <- 3000
deuda <- 500
entrada <- 15000
entrada_minima <- 50000
recolecta_datos_de_resultado(tasacion, tasa, cuota, prestamo, plazo_meses, fecha_inicial, 
      ingresos, deuda, entrada, entrada_minima)

}
